Lay down a base program
The project is designed to work with a database storing Learning Objects. Using Firebase database, objects can be stored and retrieved from the database. The Objects can be removed and updated through a few buttons. TODO: build a UI for the user and a form to allow the user to customize the Learning Objects